Rename and simplify into StreamProcessorWorkflow
